fixing sign
